moved helpers folder to utils, created constants folder, updated imports, deleted comment, fixed height of delete icon inside the resentSearches
