+++ dispatcher-client/src/components/CardSecondary/style.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { COLORS } from '../../utils/colors';
+import { COLORS } from '../../utils/constants/colors';
 import {
   CardSharedStyle,
   FlexColumn,
