+++ dispatcher-client/src/components/dashboard/components/newsWidget/styles.ts
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import { FlexColumn } from '../../../../styles/sharedStyles';
-import { COLORS } from '../../../../utils/colors';
-import { SCREENS } from '../../../../utils/screenSizes';
+import { COLORS } from '../../../../utils/constants/colors';
+import { SCREENS } from '../../../../utils/constants/screenSizes';
 
 export const CardsContainer = styled(FlexColumn)`
   overflow-y: auto;
