+++ dispatcher-client/src/components/datePicker/DatePickerComponent.tsx
@@ -5,7 +5,7 @@ import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
 import { DatePicker } from '@mui/x-date-pickers/DatePicker';
 import { ReactComponent as DateIcon } from '../../assets/svgs/date.svg';
 import { DatesFilterContainer } from './styles';
-import { COLORS } from '../../utils/colors';
+import { COLORS } from '../../utils/constants/colors';
 
 const DatePickerComponent: React.FC = () => {
   const [value, setValue] = React.useState<Date | null>(null);
