+++ dispatcher-client/src/styles/sharedStyles.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { COLORS } from '../utils/colors';
+import { COLORS } from '../utils/constants/colors';
 
 export const FlexColumn = styled.div`
   display: flex;
