+++ dispatcher-client/src/utils/helperFunctions/isRTL.ts
